Implement playRound and findRoundWinner
Commit README. Finish getPlayerChoice. Implement playRound, which
gets player and computer input, finds the round winner with
findRoundWinner, and increments score accordingly. Implement
findRoundWinner, which uses a base list of rock, paper, scissors, and
makes a new list to compare it to depending on player choice, which
has the round outcome at the same indeces for each possible compChoice
in the rpsList.