Refactor program to prepare for adding ui
Remove getPlayerChoice function because playerChoice comes from the
event listeners instead of user text input. Make playRound take
playerChoice as an argument so it can be called directly with the
choice from the eventListener function on the buttons. Make global
boolean variable gameOver that only allows buttons to be pressed if
the game is still in progress. Make function checkWinner to check if
the game is over. Remove the loop and make round a global variable.